Append line numbers to node addresses to prevent recycled memory from causing graph mangling
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/branches/tokenization@868 96dfbd92-d197-e211-9ca9-00110a534b34

+++ as6502/main.c
@@ -166,7 +166,7 @@ static void assembleFile(FILE *in, FILE *out, int printProcess, int printTable, 
 	if (printDot) {
 		printf("digraph G {\n");
 		printf("{ node [shape = plaintext]; ");
-		for (int i = 1; i < currentLineNum ; i++) {
+		for (size_t i = 1; i < currentLineNum ; i++) {
 			printf("%lu -> ", i);
 		}
 		printf("%lu; }", currentLineNum);
